refactor(packs): split manager tests for ci limit
